+++ NEWS.md
@@ -3,6 +3,8 @@
 * Header values (of version made by and external attributes) are now correctly
   read and written on big-endian systems (#68).
 
+* `zip_list()` now also returns `crc32` and `offset` (#74, thanks @jefferis).
+
 # 2.1.1
 
 This version has no user visible changes.